Refactor - Modularized the Network Monitor TG-57 #done
