+++ src/app.container.ts
@@ -3,12 +3,12 @@ import { container } from 'inversify-props'
 import IHomeDashboardService from '@/components/home-dashboard/services/IHomeDashboardService'
 import IAddonsService from '@/services/interfaces/IAddonsService'
 import ICaseFileService from '@/components/case-file/services/ICaseFileService'
-import INetworkMonitorService from '@/services/interfaces/INetworkMonitorService'
+import INetworkMonitorService from '@/components/network-monitor/services/INetworkMonitorService'
 import ICameraMonitorService from '@/services/interfaces/ICameraMonitorService'
 import AddonsService from '@/services/implementations/AddonsService'
 import CaseFileService from '@/components/case-file/services/CaseFileService'
 import HomeDashboardService from '@/components/home-dashboard/services/HomeDashboardService'
-import NetworkMonitorService from '@/services/implementations/NetworkMonitorService'
+import NetworkMonitorService from '@/components/network-monitor/services/NetworkMonitorService'
 import TYPES from '@/InjectableTypes/types'
 import IMenuItemService from '@/services/interfaces/IMenuItemService'
 import MenuItemService from '@/services/implementations/MenuItemService'
