+++ src/components/NavDrawer.vue
@@ -69,7 +69,7 @@ import IMenuItem from '@/types/IMenuItem'
 import IHomeDashboardService from '@/components/home-dashboard/services/IHomeDashboardService'
 import IAddonsService from '@/services/interfaces/IAddonsService'
 import ICaseFileService from './case-file/services/ICaseFileService'
-import INetworkMonitorService from '@/services/interfaces/INetworkMonitorService'
+import INetworkMonitorService from './network-monitor/services/INetworkMonitorService'
 import ICameraMonitorService from '@/services/interfaces/ICameraMonitorService'
 import TYPES from '@/InjectableTypes/types'
