+++ src/components/network-monitor/render/NetworkMonitorCard.vue
