+++ src/components/network-monitor/services/INetworkMonitorService.ts
@@ -1,4 +1,4 @@
-import IService from './IService'
+import IService from '@/services/interfaces/IService'
 
 export default interface INetworkMonitorService extends IService {
 }
