+++ src/components/network-monitor/services/NetworkMonitorService.ts
